WIP - implement home underwriting guidelines

+++ src/main/kotlin/com/hedvig/underwriter/serviceIntegration/productPricing/Dtos/QuotePriceResponseDto.kt
@@ -1,6 +1,8 @@
 package com.hedvig.underwriter.serviceIntegration.productPricing.Dtos
 
+import java.math.BigDecimal
+
 data class QuotePriceResponseDto (
-        val price: Double?
+    val price: BigDecimal?
 )
